Now we pass the '-l gi' (= scan-switching) iteration consistency assertion.
The order of operations in the Soft-DFS and Best-First-Search ("A-star")
scans was changed to accomodate for the iteration direct-incrementation
consistency.

Individual changes:

scans.c : made the traces run only if the FCS_TRACE environment variable
is defined, and fixed the problems.

t/t/intermediate-iterations-are-in-order.t : removed the $TODO from the 
"-l gi" assertion because it is now successful.