Two bugfixes: Off-by-1 final iters count. & iter_handler unapplied globally.
The off-by 1 final iterations count is manifested in keeping num_times
the same after a solution was reached. So it is lower by 1 than the actual
one. The number of boards checked is greater by 1 than the reported iterations
which was equivalent to the iteration number of the last iteration.

The iter_handler was caused by assigning the iter_handler only to the latest
instance in fcs_user_t and not to all of them.