Branch to the investigate-p4-slowdown-by-r2911.
This was prompted by:

http://tech.groups.yahoo.com/group/fc-solve-discuss/message/1061

[quote]
However, there was a fly in the ointment: I noticed that a previous scan "the-
iglu-cabal" ran slower on my stationary P4-2.4GHz machine than before I
implemented the "depth-tests-order". After a lot of investigation, I realised
that the problem was with the first commit in the
add-depth-dependent-moves-order branch (there was a slowdown between r2910 and
r2911). The commit seems very innocent and I just add a new command line
option there, so I don't know why there was a problem.

That put aside, there doesn't appear to be a slowdown on my x86-64 (Dual Core)
Laptop so I don't know why the P4 doesn't like it. I've used the same version
(gcc-4.5.0 with -flto) of the compiler on both systems. This is really very
frustrating because on the Pentium 4 all the gains from the new preset are
lost due to this slowdown.

I'll try to investigate further because I don't want this slowdown.
[/quote]